feat: render channels
